host_agent/pyproject.toml
[project]
name = "host_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.16",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "google-adk>=1.7.0",
    "mcp[cli]>=1.12.0",
]
